fix: merge utils and utilities / impl new error pages
