+++ inc/AutoindexHandler.hpp
@@ -2,7 +2,7 @@
 
 #include "FileSystemPolicy.hpp"
 #include "Directory.hpp"
-#include "utils.hpp"
+#include "utilities.hpp"
 
 #include <string>
 #include <vector>
