+++ test/test_utils.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "utils.hpp"
+#include "utilities.hpp"
 
 TEST(getFileExtension, NoExtension)
 {
